Move Slurm logic to plugins
